1. Consolidate `_get_color_{name, foreground, background}` into `_get_color_property` 2. Move `plist` into `lib/` 3. Move to `"".format()` from `"%s" % value` formatting 4. Fix `add_regions` when no icon is desired 5. Rename `key, value` to `index, option` for `_get_tag_flags` 6. Remove unneeded `global` settings 7. Rename `get_settings` to `load_settings`
